write out each py file instad of template string
